(+) display block table model

+++ src/gui/mainwindowcontroller.cpp
@@ -8,24 +8,28 @@
 using BBTCalculator::Core::Core;
 using BBTCalculator::Gui::MainWindowController;
 
-MainWindowController::MainWindowController() :
-    core{nullptr}
-{}
+MainWindowController::MainWindowController()
+    : core{nullptr}
+{
+}
 
-void MainWindowController::setCore(BBTCalculator::Core::Core *c)
+void MainWindowController::setCore(BBTCalculator::Core::Core* c)
 {
     core = c;
 }
 
 void MainWindowController::onOpenWorkspaceClicked(bool)
 {
-  core->letUserSelectWorkspace();
+    core->letUserSelectWorkspace();
 }
 
-void MainWindowController::onLocSelectionChanged(const QItemSelection& selected,
-                                                 const QItemSelection& deselected)
+void MainWindowController::onLocSelectionChanged(
+    const QItemSelection& selected, const QItemSelection& deselected)
 {
     const QItemSelectionRange& singleRow{selected.at(0)};
 
-    core->displayImageForLocName(singleRow.indexes().at(LocViewColumns::NAME).data().toString());
+    core->displayImageForLocName(singleRow.indexes()
+                                     .at(static_cast<int>(LocViewColumns::NAME))
+                                     .data()
+                                     .toString());
 }
